+++ components/LocusZoom/config.ts
