+++ components/Table/Cell.tsx
@@ -1,8 +1,8 @@
 import React, { ReactNode } from 'react'
 
 
-import { BasicType, Modify, TypeMapper, Expand, NAString } from "@common/types"
-import { _isJSON, _deepCopy, _hasOwnProperty, _get, _isNull, _isNA } from '@common/utils'
+import { BasicType, Modify, TypeMapper, Expand, NAString } from "@root/common/types"
+import { _isJSON, _deepCopy, _hasOwnProperty, _get, _isNull, _isNA } from '@root/common/utils'
 import { Color } from '@common/palettes'
 
 import { Text, TextList } from '@text/BasicText'
