+++ components/TextRenderers/BasicText.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react"
 
-import { _deepCopy, _get, _hasOwnProperty, _isJSON, _isNA, _isNull } from "@common/utils";
+import { _deepCopy, _get, _hasOwnProperty, _isJSON, _isNA, _isNull } from "../../../common/utils
 
 import {
     TextRenderer,
