+++ components/TextRenderers/TextRenderer.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from "react"
 
-import { _get, _hasOwnProperty, _isNA, _isNull } from "@common/utils";
+import { _get, _hasOwnProperty, _isNA, _isNull } from "../../../common/utils
 import { renderTooltip } from "@components/UI/Tooltip";
 
 import {
