+++ package.json
@@ -3,8 +3,6 @@
     "version": "0.1.0",
     "author": "fossilfriend,bug-sam",
     "scripts": {
-        "prepare": "npm run build",
-        "build": "rollup -c --bundleConfigAsCjs",
         "storybook": "storybook dev -p 6006",
         "build-storybook": "storybook build",
         "storybook-docs": "storybook dev --docs"
@@ -57,6 +55,7 @@
         "match-sorter": "^6.3.1",
         "react-error-boundary": "^4.0.13",
         "react-router-dom": "^6.19.0",
-        "ts-interface-checker": "^1.0.2"
+        "ts-interface-checker": "^1.0.2",
+        "next": "^14.1.0"
     }
 }
