--- src/components/index.ts
@@ -1,6 +1,0 @@
-//export * from "./Table"
-//export * from "./TextRenderers"
-//export * from "./Chart/Highcharts"
-//export { LocusZoomPlot, MemoLocusZoomPlot} from "./LocusZoom"
-
-export {}
