+++ stories/DataViz/Table/LargeTable.stories.ts
@@ -2,7 +2,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
 import Table from '@table/Table';
-import { TABLE_DEFINTION as table } from '../../../../example/data/table_large_numeric_values';
+import { TABLE_DEFINTION as table } from '@examples/tables/table_large_numeric_values';
 
 const meta: Meta<typeof Table> = {
     title: 'NIAGADS-VIZ/Data Vizualization/Table/Large Table',
