+++ tsconfig.build.json
@@ -10,7 +10,6 @@
         "build",
         "dist",
         "example",
-        "scripts",
-        "src/stories"
+        "stories"
     ]
 }
